fix: background remover styling , dashboard schedule navigation, UUID placeholder, and toolbox item corrections
