+++ src/app/ui/components/toolboxfolder/colorPicker/ColorPicker.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import toast from "react-hot-toast";
 import { useNavigate } from "react-router-dom";
+import ColorPickerItem from "./colorPickerItem";
 
 const ColorPicker: React.FC = () => {
     const [color, setColor] = useState("#0099e5");
@@ -86,95 +87,13 @@ const ColorPicker: React.FC = () => {
                     </p>
                 </div>
             </div>
-            <div style={styles.container}>
-
-                <h2 style={{ color: "#071D6A", fontWeight: "900" }}>Simple Color Picker</h2>
-
-                <input
-                    type="color"
-                    value={color}
-                    onChange={handleColorChange}
-                    style={{ ...styles.inputColor, backgroundColor: color }}
-                />
-
-                <div style={styles.outputGroup}>
-                    <OutputRow label="HEX" value={color} onCopy={handleCopy} />
-                    <OutputRow label="RGB" value={rgb} onCopy={handleCopy} />
-                    <OutputRow label="HSL" value={hsl} onCopy={handleCopy} />
-                </div>
-
-                <div style={{ ...styles.preview, backgroundColor: color }}>
-                    <p style={{ color: "#fff" }}>Color Preview</p>
-                </div>
-            </div>
+            <ColorPickerItem/>
         </>
     );
 };
 
-const OutputRow: React.FC<{ label: string, value: string, onCopy: (text: string) => void }> = ({ label, value, onCopy }) => (
-    <div style={{ ...styles.outputRow, color: "#000000" }}>
-        <span style={{ fontWeight: "bold" }}>{label}:</span>
-        <button
-            onClick={() => onCopy(value)}
-            style={styles.copyButton}
-            disabled={!value}
-        >
-            📋 {value}
-        </button>
-    </div>
-);
 
-const styles: { [key: string]: React.CSSProperties } = {
-    container: {
-        maxWidth: "500px",
-        margin: "2rem auto",
-        padding: "2rem",
-        backgroundColor: "#f8f9fa",
-        borderRadius: "10px",
-        textAlign: "center",
-        fontFamily: "sans-serif",
-        boxShadow: "0 4px 12px rgba(0, 0, 0, 0.1)",
-    },
-    inputColor: {
-        width: "100%",
-        height: "60px",
-        border: "none",
-        borderRadius: "8px",
-        cursor: "pointer",
-        marginBottom: "1.5rem",
-    },
-    outputGroup: {
-        display: "flex",
-        flexDirection: "column",
-        gap: "1rem",
-        marginBottom: "1.5rem",
-    },
-    outputRow: {
-        display: "flex",
-        justifyContent: "space-between",
-        alignItems: "center",
-        backgroundColor: "#fff",
-        border: "1px solid #ccc",
-        padding: "0.75rem 1rem",
-        borderRadius: "5px",
-    },
-    copyButton: {
-        backgroundColor: "#071D6A",
-        color: "#fff",
-        border: "none",
-        padding: "0.5rem 1rem",
-        borderRadius: "5px",
-        cursor: "pointer",
-        fontSize: "0.9rem",
-    },
-    preview: {
-        height: "100px",
-        borderRadius: "8px",
-        display: "flex",
-        justifyContent: "center",
-        alignItems: "center",
-        marginTop: "1rem",
-    },
-};
+
+
 
 export default ColorPicker;
