+++ src/app/ui/pages/toolboxpage/premiumtoolbox/BackgroundRemove.tsx
@@ -3,19 +3,20 @@
 import React from "react"
 import { useState, useRef, useCallback } from "react"
 import {
-  Upload,
-  X,
+
+  
   ImageIcon,
-  Download,
-  Loader2,
-  RotateCcw,
-  Settings,
-  ChevronDown,
+  
+ 
+
+  
+
   Zap,
-  Key,
-  ExternalLink,
+
+
 } from "lucide-react"
 import "./BackgroundRemover.css"
+import BackgroundRemoverItem from "./BackgroundRemoverItem"
 
 interface ProcessingState {
   isProcessing: boolean
@@ -217,34 +218,10 @@ const BackgroundRemove: React.FC<BackgroundRemoverProps> = ({ onClose }) => {
     }
   }
 
-  const downloadImage = () => {
-    if (!processedImage) return
+ 
 
-    const link = document.createElement("a")
-    link.href = processedImage
-    const formatExtension = outputFormat === "jpeg" ? "jpg" : outputFormat
-    link.download = `background-removed-${Date.now()}.${formatExtension}`
-    document.body.appendChild(link)
-    link.click()
-    document.body.removeChild(link)
-  }
 
-  const resetImages = () => {
-    setOriginalImage(null)
-    setProcessedImage(null)
-    setError(null)
-    if (fileInputRef.current) {
-      fileInputRef.current.value = ""
-    }
-  }
 
-  const saveApiKey = () => {
-    if (apiKey.trim()) {
-      localStorage.setItem("removebg_api_key", apiKey.trim())
-      setShowApiKeyInput(false)
-      setError(null)
-    }
-  }
 
   // Load API key from localStorage on component mount
   React.useEffect(() => {
@@ -273,204 +250,7 @@ const BackgroundRemove: React.FC<BackgroundRemoverProps> = ({ onClose }) => {
           )}
         </div>
       </div>
-
-      <div className="airbg-main">
-        {!originalImage ? (
-          <div
-            className={`airbg-upload-zone ${dragActive ? "airbg-drag-active" : ""}`}
-            onDragEnter={handleDrag}
-            onDragLeave={handleDrag}
-            onDragOver={handleDrag}
-            onDrop={handleDrop}
-            onClick={() => fileInputRef.current?.click()}
-          >
-            <Upload className="airbg-upload-icon" />
-            <h3 className="airbg-upload-title">Drop your image here</h3>
-            <p className="airbg-upload-text">or click to browse files</p>
-            <div className="airbg-upload-formats">
-              <span>Supports: JPG, PNG, WEBP</span>
-              <span>Max size: 12MB</span>
-            </div>
-            <div className="airbg-removebg-note">
-              <Zap size={16} />
-              <span>Powered by D'roid.bg • Advanced AI Processing</span>
-            </div>
-            <input
-              ref={fileInputRef}
-              type="file"
-              accept="image/*"
-              onChange={handleFileInput}
-              className="airbg-file-input"
-            />
-          </div>
-        ) : (
-          <div className="airbg-editor">
-            <div className="airbg-editor-header">
-              <button onClick={resetImages} className="airbg-reset-btn">
-                <RotateCcw size={16} />
-                Start Over
-              </button>
-            </div>
-
-            <div className="airbg-settings-panel">
-              <button onClick={() => setShowSettings(!showSettings)} className="airbg-settings-toggle">
-                <Settings size={16} />
-                API & Output Settings
-                <ChevronDown size={16} className={showSettings ? "rotate-180" : ""} />
-              </button>
-
-              {showSettings && (
-                <div className="airbg-settings-content">
-                  <div className="airbg-setting-group">
-                    <label className="airbg-setting-label">D'roid.bg API Key</label>
-                    {!showApiKeyInput && apiKey ? (
-                      <div className="airbg-api-key-display">
-                        <div className="airbg-api-key-masked">
-                          <Key size={16} />
-                          <span>API Key: ••••••••{apiKey.slice(-4)}</span>
-                          <button onClick={() => setShowApiKeyInput(true)} className="airbg-change-key-btn">
-                            Change
-                          </button>
-                        </div>
-                      </div>
-                    ) : (
-                      <div className="airbg-api-key-input">
-                        <input
-                          type="password"
-                          value={apiKey}
-                          onChange={(e) => setApiKey(e.target.value)}
-                          placeholder="Enter your Remove.bg API key"
-                          className="airbg-setting-input"
-                        />
-                        <button onClick={saveApiKey} className="airbg-save-key-btn">
-                          Save
-                        </button>
-                      </div>
-                    )}
-                    <div className="airbg-api-help">
-                      <ExternalLink size={14} />
-                      <span>Powered by advanced AI technology for professional results</span>
-                    </div>
-                  </div>
-
-                  <div className="airbg-setting-group">
-                    <label className="airbg-setting-label">Output Format</label>
-                    <select
-                      value={outputFormat}
-                      onChange={(e) => setOutputFormat(e.target.value as "png" | "webp" | "jpeg")}
-                      className="airbg-setting-select"
-                    >
-                      <option value="png">PNG (Lossless, Transparent)</option>
-                      <option value="webp">WebP (Modern, Smaller Size)</option>
-                      <option value="jpeg">JPEG (Smallest Size, No Transparency)</option>
-                    </select>
-                  </div>
-
-                  {(outputFormat === "webp" || outputFormat === "jpeg") && (
-                    <div className="airbg-setting-group">
-                      <label className="airbg-setting-label">Quality: {Math.round(outputQuality * 100)}%</label>
-                      <input
-                        type="range"
-                        min="0.1"
-                        max="1"
-                        step="0.1"
-                        value={outputQuality}
-                        onChange={(e) => setOutputQuality(Number.parseFloat(e.target.value))}
-                        className="airbg-setting-slider"
-                      />
-                      <div className="airbg-quality-labels">
-                        <span>Lower Size</span>
-                        <span>Higher Quality</span>
-                      </div>
-                    </div>
-                  )}
-
-                  {outputFormat === "jpeg" && (
-                    <div className="airbg-setting-note">
-                      <X size={14} />
-                      Note: JPEG format doesn't support transparency. Background will be white.
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
-
-            <div className="airbg-image-grid">
-              <div className="airbg-image-panel">
-                <div className="airbg-panel-header">
-                  <ImageIcon size={16} />
-                  <span>Original</span>
-                </div>
-                <div className="airbg-image-container">
-                  <img
-                    ref={originalImageRef}
-                    src={originalImage || "/placeholder.svg"}
-                    alt="Original"
-                    className="airbg-image"
-                    crossOrigin="anonymous"
-                  />
-                </div>
-              </div>
-
-              <div className="airbg-image-panel">
-                <div className="airbg-panel-header">
-                  <Zap size={16} />
-                  <span>Background Removed</span>
-                </div>
-                <div className="airbg-image-container airbg-transparent-bg">
-                  {processing.isProcessing ? (
-                    <div className="airbg-processing">
-                      <Loader2 className="airbg-spinner" />
-                      <div className="airbg-progress-info">
-                        <div className="airbg-progress-text">{processing.stage}</div>
-                        <div className="airbg-progress-bar">
-                          <div className="airbg-progress-fill" style={{ width: `${processing.progress}%` }} />
-                        </div>
-                        <div className="airbg-progress-percent">{Math.round(processing.progress)}%</div>
-                      </div>
-                    </div>
-                  ) : processedImage ? (
-                    <img src={processedImage || "/placeholder.svg"} alt="Processed" className="airbg-image" />
-                  ) : (
-                    <div className="airbg-placeholder">
-                      <Zap className="airbg-placeholder-icon" />
-                      <p>Click "Remove Background" to process</p>
-                      <p className="airbg-placeholder-subtext">Powered by D'roid.bg AI</p>
-                    </div>
-                  )}
-                </div>
-              </div>
-            </div>
-
-            <div className="airbg-actions">
-              {!processing.isProcessing && !processedImage && (
-                <button
-                  onClick={processImage}
-                  className="airbg-btn airbg-btn-primary"
-                  disabled={processing.isProcessing || !apiKey.trim()}
-                >
-                  <Zap size={18} />
-                  Remove Background with D'roid.bg
-                </button>
-              )}
-
-              {processedImage && (
-                <button onClick={downloadImage} className="airbg-btn airbg-btn-success">
-                  <Download size={18} />
-                  Download Result
-                </button>
-              )}
-            </div>
-          </div>
-        )}
-
-        {error && (
-          <div className="airbg-error">
-            <X size={16} />
-            {error}
-          </div>
-        )}
-      </div>
+      <BackgroundRemoverItem/>
 
       <div className="airbg-features">
         <div className="airbg-feature">
